refactor: pagination

+++ lib/apps/infra/api/stores/dto/stores_query.dart
@@ -9,8 +9,8 @@ class StoresQuery with _$StoresQuery {
   factory StoresQuery({
     required double lat,
     required double lng,
-    double? page,
-    double? perPage,
+    int? page,
+    int? perPage,
     OperationStatus? operationStatus,
     String? minOpenTime,
     String? maxOpenTime,
